[Edited by dschult: text added from email to the NetworkX-discuss list from @YVWX ]
 just made a PR (#6968). I find the current implementation of Dinitz's algorithm is actually an Edmonds-Karp algorithm, so I try to make a change.

The second issue is about the documentation. There is a shortest_augmenting_path algorithm. As I see, the core part of the algorithm is the same as Edmonds_Karp. However, its time complexity mentioned in the user guide differs from Edmonds_Karp (O(n^2m) and O(nm^2)). Could you help check this problem?

Thank you in advance! 

<!--
Please use pre-commit to lint your code.
For more details check out step 1 and 4 of
https://networkx.org/documentation/latest/developer/contribute.html
-->