@dschult 2023-10-01
This dinitz implementation was largely added via #1978 and there may be comments there which are helpful to understand the implementation. The cited paper may also help understand any difference between this implementation and the original dinitz algorithm.  Hopefully it also addresses the complexity. But I’m not sure it gives complexity for the algorithms you are wondering about.  There is also a comment there with code used to time the results (there may be better timing tools today — that was 8 years ago).  The comments there suggest that Edmonds-Karl is the faster implementation for the cases looked at. 

Does your version of the code produce the same results as the current version?  Have you done any time tests?
Why do you think the current algorithm implements Edmonds-Karp instead of Dinitz (how should I be looking for the difference between the algorithms?)