@YVWX 2023-10-03
Hi, I have uploaded my [test script](https://github.com/YVWX/TestScript/blob/main/timingtest.py). It is largely based on the work in [this thread](https://github.com/scipy/scipy/pull/10566).

The output values of my implementation are consistent with the ones of Edmonds-Karp.

Here are the timing tests (and the time is measured in seconds).

The result on the 1st graph:
```
Edmonds Karp: 0.5524605909983317
Shortest augmenting path: 0.5986862182617188
Original Dinitz: 9.085975567499796
```
The new implementation:
```
New Dinitz: 4.414525429407756
```
The result on the 2nd graph:
```
Edmonds Karp: 3.7168155511220298
Shortest augmenting path: 0.2403879165649414
Original Dinitz: 9.335507074991861
```
The new implementation:
```
New Dinitz: 1.760817050933838
```
The result on graphs of different densities.
| Algorithms and Densities | 0.1 | 0.3 | 0.5 |
| - | - | - | - |
| Edmonds Karp | 0.7506591478983561 | 2.2270760536193848| 4.318867206573486 |
| Shortest augmenting path | 0.9440836906433105 | 2.7835280895233154 | 4.686661005020142 |
| Original Dinitz | 9.547501881917318 | 60.46248229344686 | 165.1355938911438 |

The new implementation:
| 0.1 | 0.3 | 0.5 |
| - | - | - |
| 4.325359106063843 | 30.983328580856323 | 109.96503067016602 |

I find the overhead somewhat out of my expectation.

I also find the thread and the reference interesting. I figure out that my implementation matches Cherkassky's design.